refactor(cli): flatten commands into one cmd package, move flag types to pkg/cliflag
Merge the workload, definition, and flags subpackages into a single cmd
package and unexport the flag registration funcs. Rename the flagtypes
value-type package to cliflag and move it from internal/ to pkg/.
