+++ cli/cmd/flags.go
@@ -1,28 +1,26 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2026 NVIDIA Corporation
 
-// Package flags defines and registers the karta CLI's flags on cobra commands,
-// backing them with the value structures in the flagtypes package.
-package flags
+package cmd
 
 import (
 	"strings"
 
-	"github.com/run-ai/karta/cli/internal/flagtypes"
+	"github.com/run-ai/karta/cli/pkg/cliflag"
 
 	"github.com/spf13/cobra"
 )
 
-// WithKubeconfig registers the --kubeconfig persistent flag on cmd.
-func WithKubeconfig(cmd *cobra.Command) {
+// withKubeconfig registers the --kubeconfig persistent flag on cmd.
+func withKubeconfig(cmd *cobra.Command) {
 	cmd.PersistentFlags().String("kubeconfig", "",
 		"Path to the kubeconfig file to use (defaults to $KUBECONFIG or ~/.kube/config)")
 }
 
-// WithOutput registers the -o/--output enum persistent flag on cmd, backed by
-// flagtypes.Output, along with its shell completion.
-func WithOutput(cmd *cobra.Command) {
-	out := flagtypes.NewOutput()
+// withOutput registers the -o/--output enum persistent flag on cmd, backed by
+// cliflag.Output, along with its shell completion.
+func withOutput(cmd *cobra.Command) {
+	out := cliflag.NewOutput()
 	cmd.PersistentFlags().VarP(out, "output", "o",
 		"Output format: one of "+strings.Join(out.Allowed(), ", "))
 	cobra.CheckErr(cmd.RegisterFlagCompletionFunc("output",
@@ -31,9 +29,8 @@ func WithOutput(cmd *cobra.Command) {
 		}))
 }
 
-// WithNamespace registers the required -n/--namespace persistent flag on cmd
-// (single-namespace MVP).
-func WithNamespace(cmd *cobra.Command) {
+// withNamespace registers the required -n/--namespace persistent flag on cmd.
+func withNamespace(cmd *cobra.Command) {
 	cmd.PersistentFlags().StringP("namespace", "n", "",
 		"Namespace scope for workload commands (required)")
 	cobra.CheckErr(cmd.MarkPersistentFlagRequired("namespace"))
