+++ cli/pkg/cliflag/enum.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2026 NVIDIA Corporation
 
-package flagtypes
+package cliflag
 
 import (
 	"errors"
