+++ cli/pkg/cliflag/output.go
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2026 NVIDIA Corporation
 
-// Package flagtypes holds the flag value structures for the karta CLI: types
+// Package cliflag holds the flag value structures for the karta CLI: types
 // that implement pflag.Value and define a flag's shape and behavior (parse,
 // validate, print). It starts with the output-format enum.
-package flagtypes
+package cliflag
 
 // Output is the enum of supported CLI output formats.
 type Output string
